Refactor Expo native view managers
