+++ packages/expo/ios/ClerkAuthNativeView.swift
@@ -53,3 +53,15 @@ public class ClerkAuthNativeView: ClerkNativeViewHost {
     )
   }
 }
+
+@objc(ClerkAuthViewManager)
+class ClerkAuthViewManager: RCTViewManager {
+
+  override static func requiresMainQueueSetup() -> Bool {
+    return true
+  }
+
+  override func view() -> UIView! {
+    return ClerkAuthNativeView()
+  }
+}
