+++ packages/expo/ios/ClerkExpoModule.swift
@@ -6,6 +6,12 @@
 import UIKit
 import React
 
+/// Events emitted by the native view wrappers to their React Native host views.
+public enum ClerkNativeViewEvent: String {
+  /// Emitted by the Expo host view when app-owned dismissible content leaves the window.
+  case dismissed
+}
+
 // Global registry for the app-target native bridge (set by the app target at startup)
 public var clerkNativeBridge: ClerkNativeBridgeProtocol?
 
@@ -126,4 +132,9 @@ class ClerkExpoModule: RCTEventEmitter {
     }
   }
 
+}
+
+/// Requests that ClerkProvider reload the JS client from native client state.
+public func emitClerkNativeRefreshClient() {
+  ClerkExpoModule.emitRefreshClient()
 }
