+++ packages/expo/ios/ClerkUserProfileNativeView.swift
@@ -42,3 +42,15 @@ public class ClerkUserProfileNativeView: ClerkNativeViewHost {
     )
   }
 }
+
+@objc(ClerkUserProfileViewManager)
+class ClerkUserProfileViewManager: RCTViewManager {
+
+  override static func requiresMainQueueSetup() -> Bool {
+    return true
+  }
+
+  override func view() -> UIView! {
+    return ClerkUserProfileNativeView()
+  }
+}
